+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+- Changed `webhookclient.send() -> webhookclient.sendText()`
+- Add `webhookclient.send(String content, List<Map<String, dynamic>> embeds)`
+- Add `webhookclient.edit(String message-id, {String content, List<Map<String, dynamic>> embeds})`
+- Add `webhookclient.editText(String message-id, String content)`
+- Add `webhookclient.get(String message-id)`
+- Add `webhookclient.delete(String message-id)`
+
 ## 1.0.0
 
 - Initial version.